chore: use pooler profiles
