+++ pkg/management/postgres/utils/roles.go
@@ -39,25 +39,30 @@ func DisableSuperuserPassword(db *sql.DB) error {
 		return nil
 	}
 
-	return ExecuteWithLocalCommit(db, func(tx *sql.Tx) error {
-		_, err = tx.Exec("ALTER ROLE postgres WITH PASSWORD NULL")
-		if err != nil {
-			return fmt.Errorf("while running ALTER ROLE %v WITH PASSWORD: %w", "postgres", err)
-		}
+	tx, err := db.Begin()
+	if err != nil {
+		return err
+	}
+	defer func() {
+		// This has no effect if the transaction
+		// is committed
+		_ = tx.Rollback()
+	}()
 
-		return nil
-	})
+	// we don't want to be stuck here if synchronous replicas are still not alive
+	// and kicking
+	_, err = tx.Exec("ALTER ROLE postgres WITH PASSWORD NULL")
+	if err != nil {
+		return fmt.Errorf("while running ALTER ROLE %v WITH PASSWORD: %w", "postgres", err)
+	}
+
+	return tx.Commit()
 }
 
 // SetUserPassword change the password of a user in the PostgreSQL database
 func SetUserPassword(username string, password string, db *sql.DB) error {
-	return ExecuteWithLocalCommit(db, func(tx *sql.Tx) error {
-		_, err := tx.Exec(fmt.Sprintf("ALTER ROLE %v WITH PASSWORD %v",
-			pgx.Identifier{username}.Sanitize(),
-			pq.QuoteLiteral(password)))
-		if err != nil {
-			return fmt.Errorf("while running ALTER ROLE %v WITH PASSWORD: %w", username, err)
-		}
-		return nil
-	})
+	_, err := db.Exec(fmt.Sprintf("ALTER ROLE %v WITH PASSWORD %v",
+		pgx.Identifier{username}.Sanitize(),
+		pq.QuoteLiteral(password)))
+	return err
 }
