+++ pkg/management/postgres/utils/roles_test.go
@@ -67,8 +67,6 @@ var _ = Describe("Credentials management functions", func() {
 		FROM pg_catalog.pg_authid
 		WHERE rolname='postgres'`).WillReturnRows(rowsHasPassword)
 		mock.ExpectBegin()
-		mock.ExpectExec("SET LOCAL synchronous_commit to LOCAL").
-			WillReturnResult(sqlmock.NewResult(0, 0))
 		mock.ExpectExec("ALTER ROLE postgres WITH PASSWORD NULL").
 			WillReturnResult(sqlmock.NewResult(0, 1))
 		mock.ExpectCommit()
@@ -77,24 +75,14 @@ var _ = Describe("Credentials management functions", func() {
 	})
 
 	It("can set the password for a PostgreSQL role", func() {
-		mock.ExpectBegin()
-		mock.ExpectExec("SET LOCAL synchronous_commit to LOCAL").
-			WillReturnResult(sqlmock.NewResult(0, 0))
 		mock.ExpectExec("ALTER ROLE \"testuser\" WITH PASSWORD 'testpassword'").
 			WillReturnResult(sqlmock.NewResult(0, 0))
-		mock.ExpectCommit()
-
 		Expect(SetUserPassword("testuser", "testpassword", db)).To(Succeed())
 	})
 
 	It("will correctly escape the password if needed", func() {
-		mock.ExpectBegin()
-		mock.ExpectExec("SET LOCAL synchronous_commit to LOCAL").
-			WillReturnResult(sqlmock.NewResult(0, 0))
 		mock.ExpectExec("ALTER ROLE \"testuser\" WITH PASSWORD 'this \"is\" weird but ''possible'''").
 			WillReturnResult(sqlmock.NewResult(0, 0))
-		mock.ExpectCommit()
-
 		Expect(SetUserPassword("testuser", "this \"is\" weird but 'possible'", db)).To(Succeed())
 	})
 })
